Made missiles explode when they lose their target...I have no idea why they're chasing Player_Attacks

+++ core/src/main/swapship/components/other/SpawnerComp.java
@@ -4,7 +4,7 @@ import main.swapship.components.TimerComp;
 
 public class SpawnerComp extends TimerComp {
 
-	public void setValues(int spawnRate) {
+	public void setValues(float spawnRate) {
 		super.startTime = spawnRate;
 		super.timeLeft = 0;
 	}
